requirements.txt
django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytest~=7.3.1
drf-spectacular==0.26.2